+++ demo/nextjs/pages/highlight-trigger-none/index.tsx
@@ -1,0 +1,72 @@
+import * as React from 'react';
+import { Viewer, Worker } from '@react-pdf-viewer/core';
+import { defaultLayoutPlugin } from '@react-pdf-viewer/default-layout';
+import { highlightPlugin, Trigger } from '@react-pdf-viewer/highlight';
+
+import type { RenderHighlightsProps } from '@react-pdf-viewer/highlight';
+
+const IndexPage = () => {
+    const defaultLayoutPluginInstance = defaultLayoutPlugin();
+
+    const areas = [
+        {
+            pageIndex: 3,
+            height: 1.55401,
+            width: 28.1674,
+            left: 27.5399,
+            top: 15.0772,
+        },
+        {
+            pageIndex: 3,
+            height: 1.32637,
+            width: 37.477,
+            left: 55.7062,
+            top: 15.2715,
+        },
+        {
+            pageIndex: 3,
+            height: 1.55401,
+            width: 28.7437,
+            left: 16.3638,
+            top: 16.6616,
+        },
+    ];
+
+    const renderHighlights = (props: RenderHighlightsProps) => (
+        <div>
+            {areas
+                .filter((area) => area.pageIndex === props.pageIndex)
+                .map((area, idx) => (
+                    <div
+                        key={idx}
+                        style={Object.assign(
+                            {},
+                            {
+                                background: 'yellow',
+                                opacity: 0.4,
+                            },
+                            props.getCssProperties(area, props.rotation)
+                        )}
+                    />
+                ))}
+        </div>
+    );
+
+    const highlightPluginInstance = highlightPlugin({
+        renderHighlights,
+        trigger: Trigger.None,
+    });
+
+    return (
+        <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.9.359/build/pdf.worker.js">
+            <div className="demo">
+                <Viewer
+                    fileUrl="/pdf-open-parameters.pdf"
+                    plugins={[defaultLayoutPluginInstance, highlightPluginInstance]}
+                />
+            </div>
+        </Worker>
+    );
+};
+
+export default IndexPage;